fixed I18n pluralization in case of foreign language and introduced options to not downcase entry_name, if requested
