+++ lib/kaminari/helpers/action_view_extension.rb
@@ -88,7 +88,10 @@ module Kaminari
     #   <%= page_entries_info @posts, :entry_name => 'item' %>
     #   #-> Displaying items 6 - 10 of 26 in total
     def page_entries_info(collection, options = {})
-      entry_name = options[:entry_name] || collection.entry_name(options.merge(count: collection.total_count))
+      total_count = collection.total_count == 1 ? 1 : 2
+      default_entry_name = options[:entry_name]
+      default_entry_name = default_entry_name.pluralize unless default_entry_name.blank? || collection.total_count == 1
+      entry_name = default_entry_name || collection.entry_name(options.merge(:count => total_count))
 
       if collection.total_pages < 2
         t('helpers.page_entries_info.one_page.display_entries', :entry_name => entry_name, :count => collection.total_count)
